# apps/api/openapi.v1.yaml
openapi: 3.1.0
info:
  title: EVE Market Platform API
  version: 1.0.0
servers:
  - url: http://localhost:3001
paths:
  /health:
    get:
      summary: Liveness check
      responses:
        '200':
          description: OK
  /ready:
    get:
      summary: Readiness check (DB)
      responses:
        '200':
          description: Ready
        '503':
          description: Not ready
  /metrics:
    get:
      summary: Metricas de aplicacion (Prometheus text format)
      responses:
        '200':
          description: OK
  /v1/regions:
    get:
      summary: Lista regiones EVE
      responses:
        '200':
          description: OK
  /v1/regions/{regionId}/constellations:
    get:
      summary: Lista constelaciones por region
      parameters:
        - in: path
          name: regionId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
  /v1/constellations/{constellationId}/stations:
    get:
      summary: Lista estaciones NPC por constelacion
      parameters:
        - in: path
          name: constellationId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
  /v1/market/snapshots:
    get:
      summary: Lista snapshots persistidos
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - in: query
          name: offset
          required: false
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: OK
    post:
      summary: Crea snapshot de mercado filtrado por estaciones NPC de constelacion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [regionId, constellationId]
              properties:
                regionId:
                  type: integer
                constellationId:
                  type: integer
      responses:
        '200':
          description: Snapshot creado
  /v1/market/snapshots/{snapshotId}:
    get:
      summary: Obtiene snapshot persistido
      parameters:
        - in: path
          name: snapshotId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '404':
          description: No encontrado
  /v1/cache/refresh:
    post:
      summary: Limpia cache backend de ESI
      responses:
        '200':
          description: OK
  /v1/logs/esi:
    post:
      summary: Ingesta logs estructurados de frontend/backend
      responses:
        '202':
          description: Accepted
  /v1/esi/{path}:
    get:
      summary: Proxy ESI (compatibilidad incremental)
      parameters:
        - in: path
          name: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
    post:
      summary: Proxy ESI POST (compatibilidad incremental)
      parameters:
        - in: path
          name: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK